backend/requirements.txt
flask
flask_sqlalchemy
flask_cors
psycopg2-binary
python-dotenv
PyJWT
